Fix Bug Update Profile

+++ app/Http/Controllers/ProfileController.php
@@ -34,8 +34,7 @@ class ProfileController extends Controller
             $request->user()->email_verified_at = null;
         }
 
-
-        if ($request->avatar) {
+        if ($request->avatar && Str::contains($request->avatar, 'tmp')) {
             if (!empty($request->user()->avatar)) {
                 Storage::disk(config('filesystems.default_public_disk'))->delete($request->user()->avatar);
             }
